requirements.txt
mcp[cli]
httpx
streamlit
groq
python-dotenv
